requirements.txt
haikunator==2.1.0
channels==1.1.3
Django==1.11.1
asgiref==1.1.1
dj_database_url==0.4.2
pytest==3.0.7
whitenoise
psycopg2
asgi_redis
djangorestframework
drf-nested-routers
djangorestframework-jwt
django-cors-headers